# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitness AI API Documentation",
        "description": "L5 Swagger OpenApi description for the Fitness Backend",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/v1/users/{username}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Perfil público do usuário",
                "description": "Retorna os dados públicos de um usuário pelo username.",
                "operationId": "321179894806795d27c295cbd761d90c",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "Username do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do perfil público"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{username}/achievements": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Conquistas do usuário",
                "description": "Retorna as conquistas desbloqueadas de um usuário.",
                "operationId": "fa9f8fd5f7b1a5db26d6998eb3884d10",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "Username do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de conquistas"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{username}/goals": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Metas do usuário",
                "description": "Retorna as metas públicas de um usuário.",
                "operationId": "b0bf4856fd64bf877e4fa05298cacdf0",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "Username do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados das metas"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/search": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Buscar usuários",
                "description": "Pesquisa usuários por nome ou username. Exclui usuários bloqueados.",
                "operationId": "d697607c7cd85fb8e5fc36bfdf9a93f8",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Termo de busca (mínimo 3 caracteres)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de usuários encontrados"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/workouts/finish": {
            "post": {
                "tags": [
                    "Workouts"
                ],
                "summary": "Salva e analisa um treino finalizado",
                "description": "Salva o log do treino, os exercícios realizados, e envia os dados para a IA (Gemini) calcular calorias gastas, músculos treinados e gerar um feedback motivacional.",
                "operationId": "b4c89b08fd7ce583d5962e421278492f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-15"
                                    },
                                    "time_start": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "14:00:00"
                                    },
                                    "time_end": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "15:30:00"
                                    },
                                    "plan_workout_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "uuid_here"
                                    },
                                    "observations": {
                                        "type": "string",
                                        "example": "Treino muito focado, mas ombro doeu um pouco."
                                    },
                                    "exercises": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "exercise_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "sets": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "reps": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "weight_kg": {
                                                    "type": "number",
                                                    "example": 20.5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Treino salvo e analisado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação ou erro na integração com a IA"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}